docs: update permissions listing and help
