+++ src/main/java/com/daveestar/bettervanilla/enums/Permissions.java
@@ -11,6 +11,7 @@ public enum Permissions {
   TIMER("bettervanilla.timer"),
   ADMINHELP("bettervanilla.adminhelp"),
   SETTINGS("bettervanilla.settings"),
+  ADMINSETTINGS("bettervanilla.adminsettings"),
   TOGGLELOCATION("bettervanilla.togglelocation"),
   TOGGLECOMPASS("bettervanilla.togglecompass"),
   LASTDEATH("bettervanilla.deathpoints"),
