+++ src/main/java/com/daveestar/bettervanilla/gui/SettingsGUI.java
@@ -1,359 +1,182 @@
 package com.daveestar.bettervanilla.gui;
 
-import java.util.Arrays;
-import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
 import org.bukkit.Material;
-import org.bukkit.Sound;
+import org.bukkit.block.Biome;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 
 import com.daveestar.bettervanilla.Main;
-import com.daveestar.bettervanilla.manager.AFKManager;
-import com.daveestar.bettervanilla.manager.MaintenanceManager;
+import com.daveestar.bettervanilla.manager.CompassManager;
+import com.daveestar.bettervanilla.manager.NavigationManager;
 import com.daveestar.bettervanilla.manager.SettingsManager;
+import com.daveestar.bettervanilla.utils.ActionBar;
 import com.daveestar.bettervanilla.utils.CustomGUI;
 
-import io.papermc.paper.event.player.AsyncChatEvent;
 import net.kyori.adventure.text.Component;
-import net.kyori.adventure.text.TextComponent;
 import net.md_5.bungee.api.ChatColor;
 
-public class SettingsGUI implements Listener {
+public class SettingsGUI {
   private final Main _plugin;
   private final SettingsManager _settingsManager;
-  private final AFKManager _afkManager;
-  private final MaintenanceManager _maintenanceManager;
-  private final Map<UUID, Boolean> _afkTimePending;
-  private final Map<UUID, Boolean> _maintenanceMessagePending;
+  private final NavigationManager _navigationManager;
+  private final CompassManager _compassManager;
+  private final ActionBar _actionBar;
+  private final AdminSettingsGUI _adminSettingsGUI;
 
   public SettingsGUI() {
     _plugin = Main.getInstance();
     _settingsManager = _plugin.getSettingsManager();
-    _afkManager = _plugin.getAFKManager();
-    _maintenanceManager = _plugin.getMaintenanceManager();
-    _afkTimePending = new HashMap<>();
-    _maintenanceMessagePending = new HashMap<>();
-    _plugin.getServer().getPluginManager().registerEvents(this, _plugin);
+    _navigationManager = _plugin.getNavigationManager();
+    _compassManager = _plugin.getCompassManager();
+    _actionBar = _plugin.getActionBar();
+    _adminSettingsGUI = new AdminSettingsGUI();
   }
 
   public void displayGUI(Player p) {
+    boolean isAdmin = p.hasPermission("bettervanilla.adminsettings");
+    int rows = isAdmin ? 4 : 3;
+
     Map<String, ItemStack> entries = new HashMap<>();
-    entries.put("maintenance", _createMaintenanceItem());
-    entries.put("creeperdamage", _createCreeperDamageItem());
-    entries.put("enableend", _createEnableEndItem());
-    entries.put("enablenether", _createEnableNetherItem());
-    entries.put("sleepingrain", _createSleepingRainItem());
-    entries.put("afkprotection", _createAFKProtectionItem());
-    entries.put("afktime", _createAFKTimeItem());
+    entries.put("togglelocation", _createToggleLocationItem(p));
+    entries.put("togglecompass", _createToggleCompassItem(p));
+    if (isAdmin) {
+      entries.put("adminsettings", _createAdminSettingsItem());
+    }
 
     Map<String, Integer> customSlots = new HashMap<>();
-    // first row
-    customSlots.put("maintenance", 0);
-    customSlots.put("creeperdamage", 2);
-    customSlots.put("enableend", 4);
-    customSlots.put("enablenether", 6);
-    customSlots.put("sleepingrain", 8);
-
-    // second row
-    customSlots.put("afkprotection", 12);
-    customSlots.put("afktime", 14);
+    customSlots.put("togglelocation", 2);
+    customSlots.put("togglecompass", 6);
+    if (isAdmin) {
+      customSlots.put("adminsettings", rows * 9 - 10);
+    }
 
     CustomGUI gui = new CustomGUI(_plugin, p,
         ChatColor.YELLOW + "" + ChatColor.BOLD + "» Settings",
-        entries, 3, customSlots, null,
+        entries, rows, customSlots, null,
         EnumSet.of(CustomGUI.Option.DISABLE_PAGE_BUTTON));
 
     Map<String, CustomGUI.ClickAction> actions = new HashMap<>();
-    actions.put("maintenance", new CustomGUI.ClickAction() {
+    actions.put("togglelocation", new CustomGUI.ClickAction() {
       @Override
       public void onLeftClick(Player player) {
-        _toggleMaintenance(player, null);
-        displayGUI(player);
-      }
-
-      @Override
-      public void onRightClick(Player player) {
-        if (!_maintenanceManager.getState()) {
-          player.sendMessage(Main.getPrefix() + "Enter maintenance message:");
-          _maintenanceMessagePending.put(player.getUniqueId(), true);
-          player.closeInventory();
-        } else {
-          _toggleMaintenance(player, null);
-          displayGUI(player);
+        if (!player.hasPermission("bettervanilla.togglelocation")) {
+          player.playSound(player, org.bukkit.Sound.ENTITY_VILLAGER_NO, 0.5F, 1);
+          return;
         }
-      }
-    });
-
-    actions.put("creeperdamage", new CustomGUI.ClickAction() {
-      @Override
-      public void onLeftClick(Player player) {
-        _toggleCreeperDamage(player);
+        _toggleLocation(player);
         displayGUI(player);
       }
     });
 
-    actions.put("enableend", new CustomGUI.ClickAction() {
+    actions.put("togglecompass", new CustomGUI.ClickAction() {
       @Override
       public void onLeftClick(Player player) {
-        _toggleEnd(player);
+        if (!player.hasPermission("bettervanilla.togglecompass")) {
+          player.playSound(player, org.bukkit.Sound.ENTITY_VILLAGER_NO, 0.5F, 1);
+          return;
+        }
+        _toggleCompass(player);
         displayGUI(player);
       }
     });
 
-    actions.put("enablenether", new CustomGUI.ClickAction() {
-      @Override
-      public void onLeftClick(Player player) {
-        _toggleNether(player);
-        displayGUI(player);
-      }
-    });
-
-    actions.put("sleepingrain", new CustomGUI.ClickAction() {
-      @Override
-      public void onLeftClick(Player player) {
-        _toggleSleepingRain(player);
-        displayGUI(player);
-      }
-    });
-
-    actions.put("afkprotection", new CustomGUI.ClickAction() {
-      @Override
-      public void onLeftClick(Player player) {
-        _toggleAFKProtection(player);
-        displayGUI(player);
-      }
-    });
-
-    actions.put("afktime", new CustomGUI.ClickAction() {
-      @Override
-      public void onLeftClick(Player player) {
-        player.sendMessage(Main.getPrefix() + "Enter AFK time in minutes:");
-        _afkTimePending.put(player.getUniqueId(), true);
-        player.closeInventory();
-      }
-    });
+    if (isAdmin) {
+      actions.put("adminsettings", new CustomGUI.ClickAction() {
+        @Override
+        public void onLeftClick(Player player) {
+          _adminSettingsGUI.displayGUI(player, gui);
+        }
+      });
+    }
 
     gui.setClickActions(actions);
     gui.open(p);
   }
 
-  private ItemStack _createMaintenanceItem() {
-    boolean state = _maintenanceManager.getState();
-    String message = _settingsManager.getMaintenanceMessage();
-    ItemStack item = new ItemStack(Material.IRON_BARS);
+  private ItemStack _createToggleLocationItem(Player p) {
+    boolean state = _settingsManager.getToggleLocation(p);
+    ItemStack item = new ItemStack(Material.FILLED_MAP);
     ItemMeta meta = item.getItemMeta();
     if (meta != null) {
-      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Maintenance"));
-
-      var lore = new ArrayList<String>();
-      lore.add("");
-      lore.add(ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: "
-          + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"));
-      if (message != null && !message.isEmpty()) {
-        lore.add(ChatColor.YELLOW + "» " + ChatColor.GRAY + "Message: " + ChatColor.YELLOW + message);
-      }
-      lore.add(ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Toggle");
-      lore.add(ChatColor.YELLOW + "» " + ChatColor.GRAY + "Right-Click: Toggle with message");
-
-      meta.lore(lore.stream().map(Component::text).collect(Collectors.toList()));
-      item.setItemMeta(meta);
-    }
-    return item;
-  }
-
-  private ItemStack _createCreeperDamageItem() {
-    boolean state = _settingsManager.getToggleCreeperDamage();
-    ItemStack item = new ItemStack(Material.CREEPER_HEAD);
-    ItemMeta meta = item.getItemMeta();
-    if (meta != null) {
-      meta.displayName(
-          Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Creeper Damage"));
-      meta.lore(Arrays.asList(
+      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Action-Bar Location"));
+      meta.lore(java.util.Arrays.asList(
           "",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: "
-              + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
+          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: " + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
           ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Toggle")
-          .stream().map(Component::text).collect(Collectors.toList()));
+          .stream().map(Component::text).toList());
       item.setItemMeta(meta);
     }
     return item;
   }
 
-  private ItemStack _createEnableEndItem() {
-    boolean state = _settingsManager.getEnableEnd();
-    ItemStack item = new ItemStack(Material.ENDER_EYE);
+  private ItemStack _createToggleCompassItem(Player p) {
+    boolean state = _compassManager.checkPlayerActiveCompass(p);
+    ItemStack item = new ItemStack(Material.COMPASS);
     ItemMeta meta = item.getItemMeta();
     if (meta != null) {
-      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Enable End"));
-      meta.lore(Arrays.asList(
+      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Bossbar Compass"));
+      meta.lore(java.util.Arrays.asList(
           "",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: "
-              + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
+          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: " + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
           ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Toggle")
-          .stream().map(Component::text).collect(Collectors.toList()));
+          .stream().map(Component::text).toList());
       item.setItemMeta(meta);
     }
     return item;
   }
 
-  private ItemStack _createEnableNetherItem() {
-    boolean state = _settingsManager.getEnableNether();
-    ItemStack item = new ItemStack(Material.BLAZE_ROD);
+  private ItemStack _createAdminSettingsItem() {
+    ItemStack item = new ItemStack(Material.REDSTONE_TORCH);
     ItemMeta meta = item.getItemMeta();
     if (meta != null) {
-      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Enable Nether"));
-      meta.lore(Arrays.asList(
+      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Admin Settings"));
+      meta.lore(java.util.Arrays.asList(
           "",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: "
-              + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Toggle")
-          .stream().map(Component::text).collect(Collectors.toList()));
+          ChatColor.YELLOW + "» " + ChatColor.GRAY + "Open global settings")
+          .stream().map(Component::text).toList());
       item.setItemMeta(meta);
     }
     return item;
   }
 
-  private ItemStack _createSleepingRainItem() {
-    boolean state = _settingsManager.getSleepingRain();
-    ItemStack item = new ItemStack(Material.BLUE_BED);
-    ItemMeta meta = item.getItemMeta();
-    if (meta != null) {
-      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "Sleeping Rain"));
-      meta.lore(Arrays.asList(
-          "",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: "
-              + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Toggle")
-          .stream().map(Component::text).collect(Collectors.toList()));
-      item.setItemMeta(meta);
+  private void _toggleLocation(Player p) {
+    boolean newState;
+    if (_settingsManager.getToggleLocation(p)) {
+      _settingsManager.setToggleLocation(p, false);
+      _actionBar.removeActionBar(p);
+      newState = false;
+    } else {
+      _navigationManager.stopNavigation(p);
+      _settingsManager.setToggleLocation(p, true);
+      Biome playerBiome = p.getWorld().getBiome(p.getLocation().toBlockLocation());
+      String locationText = ChatColor.YELLOW + "X: " + ChatColor.GRAY + p.getLocation().toBlockLocation().getBlockX()
+          + ChatColor.YELLOW + " Y: " + ChatColor.GRAY + p.getLocation().toBlockLocation().getBlockY()
+          + ChatColor.YELLOW + " Z: " + ChatColor.GRAY + p.getLocation().toBlockLocation().getBlockZ() + ChatColor.RED
+          + ChatColor.BOLD + " » " + ChatColor.GRAY + playerBiome.getKey();
+      _actionBar.sendActionBar(p, locationText);
+      newState = true;
     }
-    return item;
+    String stateText = newState ? "ENABLED" : "DISABLED";
+    p.sendMessage(Main.getPrefix() + "Action-Bar location is now " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
   }
 
-  private ItemStack _createAFKProtectionItem() {
-    boolean state = _settingsManager.getAFKProtection();
-    ItemStack item = new ItemStack(Material.TOTEM_OF_UNDYING);
-    ItemMeta meta = item.getItemMeta();
-    if (meta != null) {
-      meta.displayName(
-          Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "AFK Protection"));
-      meta.lore(Arrays.asList(
-          "",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "State: "
-              + (state ? ChatColor.GREEN + "ENABLED" : ChatColor.RED + "DISABLED"),
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Toggle")
-          .stream().map(Component::text).collect(Collectors.toList()));
-      item.setItemMeta(meta);
+  private void _toggleCompass(Player p) {
+    boolean newState;
+    if (_compassManager.checkPlayerActiveCompass(p)) {
+      _compassManager.removePlayerFromCompass(p);
+      _settingsManager.setToggleCompass(p, false);
+      newState = false;
+    } else {
+      _compassManager.addPlayerToCompass(p);
+      _settingsManager.setToggleCompass(p, true);
+      newState = true;
     }
-    return item;
-  }
-
-  private ItemStack _createAFKTimeItem() {
-    int minutes = _settingsManager.getAFKTime();
-    ItemStack item = new ItemStack(Material.CLOCK);
-    ItemMeta meta = item.getItemMeta();
-    if (meta != null) {
-      meta.displayName(Component.text(ChatColor.RED + "" + ChatColor.BOLD + "» " + ChatColor.YELLOW + "AFK Time"));
-      meta.lore(Arrays.asList(
-          "",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "Current: " + ChatColor.YELLOW + minutes + ChatColor.GRAY
-              + " minutes",
-          ChatColor.YELLOW + "» " + ChatColor.GRAY + "Left-Click: Set value")
-          .stream().map(Component::text).collect(Collectors.toList()));
-      item.setItemMeta(meta);
-    }
-    return item;
-  }
-
-  @EventHandler
-  public void onPlayerChat(AsyncChatEvent e) {
-    Player p = e.getPlayer();
-    UUID id = p.getUniqueId();
-    if (_afkTimePending.containsKey(id)) {
-      e.setCancelled(true);
-      String content = ((TextComponent) e.message()).content();
-      try {
-        int minutes = Integer.parseInt(content);
-        _plugin.getServer().getScheduler().runTask(_plugin, () -> {
-          _settingsManager.setAFKTime(minutes);
-          p.sendMessage(
-              Main.getPrefix() + "AFK time set to: " + ChatColor.YELLOW + minutes + ChatColor.GRAY + " minutes");
-          p.playSound(p, Sound.ENTITY_PLAYER_LEVELUP, 0.5F, 1);
-          _afkTimePending.remove(id);
-          displayGUI(p);
-        });
-      } catch (NumberFormatException ex) {
-        p.sendMessage(Main.getPrefix() + ChatColor.RED + "Please provide a valid number.");
-      }
-      return;
-    }
-
-    if (_maintenanceMessagePending.containsKey(id)) {
-      e.setCancelled(true);
-      String message = ((TextComponent) e.message()).content();
-      _plugin.getServer().getScheduler().runTask(_plugin, () -> {
-        _maintenanceMessagePending.remove(id);
-        _toggleMaintenance(p, message);
-        p.playSound(p, Sound.ENTITY_PLAYER_LEVELUP, 0.5F, 1);
-        displayGUI(p);
-      });
-    }
-  }
-
-  private void _toggleMaintenance(Player p, String message) {
-    boolean newState = !_maintenanceManager.getState();
-    _maintenanceManager.setState(newState, newState ? message : null);
     String stateText = newState ? "ENABLED" : "DISABLED";
-    p.sendMessage(
-        Main.getPrefix() + "The maintenance mode is now turned: " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
-    if (newState && message != null) {
-      p.sendMessage(Main.getPrefix() + "Message was set to: " + ChatColor.YELLOW + message);
-    }
-    _maintenanceManager.kickAll(_plugin.getServer().getOnlinePlayers());
-  }
-
-  private void _toggleCreeperDamage(Player p) {
-    boolean newState = !_settingsManager.getToggleCreeperDamage();
-    _settingsManager.setToggleCreeperDamage(newState);
-    String stateText = newState ? "ENABLED" : "DISABLED";
-    p.sendMessage(Main.getPrefix() + "Creeper damage is now turned: " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
-  }
-
-  private void _toggleEnd(Player p) {
-    boolean newState = !_settingsManager.getEnableEnd();
-    _settingsManager.setEnableEnd(newState);
-    String stateText = newState ? "ENABLED" : "DISABLED";
-    p.sendMessage(Main.getPrefix() + "The End is now " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
-  }
-
-  private void _toggleNether(Player p) {
-    boolean newState = !_settingsManager.getEnableNether();
-    _settingsManager.setEnableNether(newState);
-    String stateText = newState ? "ENABLED" : "DISABLED";
-    p.sendMessage(Main.getPrefix() + "The Nether is now " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
-  }
-
-  private void _toggleSleepingRain(Player p) {
-    boolean newState = !_settingsManager.getSleepingRain();
-    _settingsManager.setSleepingRain(newState);
-    String stateText = newState ? "ENABLED" : "DISABLED";
-    p.sendMessage(Main.getPrefix() + "Sleeping Rain is now turned: " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
-  }
-
-  private void _toggleAFKProtection(Player p) {
-    boolean newState = !_settingsManager.getAFKProtection();
-    _settingsManager.setAFKProtection(newState);
-    _afkManager.applyProtectionToAFKPlayers(newState);
-    String stateText = newState ? "ENABLED" : "DISABLED";
-    p.sendMessage(Main.getPrefix() + "AFK protection is now " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
+    p.sendMessage(Main.getPrefix() + "Bossbar compass is now " + ChatColor.YELLOW + ChatColor.BOLD + stateText);
   }
 }
