+++ src/main/java/com/daveestar/bettervanilla/manager/CompassManager.java
@@ -74,6 +74,16 @@ public class CompassManager {
     _settingsManager = _plugin.getSettingsManager();
   }
 
+  public void onPlayerJoined(Player p) {
+    if (_settingsManager.getToggleCompass(p)) {
+      _plugin.getServer().getScheduler().runTask(_plugin, () -> addPlayerToCompass(p));
+    }
+  }
+
+  public void onPlayerLeft(Player p) {
+    removePlayerFromCompass(p);
+  }
+
   public void destroy() {
     _activeCompass.values().forEach(BossBar::removeAll);
     _activeCompass.clear();
